backend/pyproject.toml
[project]
name = "gigachat-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.4",
    "gigachat>=0.1.42.post2",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-gigachat>=0.3.12",
    "pydantic-settings>=2.11.0",
    "sse-starlette>=3.0.3",
]
